Established the possibility to treat complex-valued material parameters (piezo!).
Therefore the following changes were neccessary:

We have a new type-definition in environmet.hh, the piezoMaterialType. This type consists of only two entries: realMaterialParameter and imagMaterialParamter.
If, by specification in an xml-file, piezoDataType = imagMaterialParameter (member of basePDE) was set, then during the routine defineIntegrators (piezoPDE.cc) a second stiff- and damping- bilinearform will be created and added. These will handle the imaginary parts of the stiffness and damping matrix in the piezo-equations in the sequel.
By default we consider piezoMaterialType equal to realMaterialParameter. So, do not worry for any existing examples.


The material parameters for the imaginary parts are stored in Matrix<Double> piezoMatrixC (see MaterialData). According set- and getEntry-methods can be used.
! By default the matrix piezoMatrixC contains the same values as piezoMatrix (we/I still need to modify the readPiezo method!).