Fixed the problem with history file and GSI:
The problem had the same reason as some time ago the problem with history files
and GMV. Currently we expect the constructor of the WriteResultsXYZ class to
also take care of handling the call to InitHistoryFiles();

The constructors of WriteResultsDatabase and WriteResultsGSI now both do this.

Note, however, that WriteResultsDatabase chose to over-write InitHistoryFiles()
and silently do nothing! Change it to at least issue a warning.