CFS++ lacks a certain robustness when handling Dirichlet boundary conditions. If the input mesh contains a node/dof pair that belongs to both, a region with homogeneous and a region with inhomogeneous boundary conditions, this will result in SinglePDE::SetBCs() passing a zero equation number down to OLAS' SetDirichlet() method, resulting in a segmentation violation.
As a temporary fix for this, I have added a check in the inhom. Dirichlet
part of SetBCs() that checks for this occurence and issues a corresponding
error message.

We will have to discuss, how we treat this situation more sensibly. Probably
the mesh reader the BCs class should already catch a problematic situation
like this one.