Added the possibility to specify the path to the directory containing the main XML schema definition file on the command line. Thus, when the schema definition changes we need not necessarily have to re-compile.
Since the path resp. XMLSCHEMA path macro is also used to locate the defaults
file, we currently really specify the path and not the filename itself, which
still must be called CFS.xsd No major headache probably.

The new parameter marker pair is:   -s / --schemaPath

As a consequence the parameter marker pair for the only-write-skeleton
flag now is:  -w / --writeSkeleton