Made the frequency-dependent adjustment of the damping parameters more flexible. So far it was assumed that the damping parameters given in the material data file are the ones for the starting frequency of the simulation.
If this is not the case, one can now specify the frequency for the material
data by setting the 'matDataFreq' element of the 'harmonics' section of the
XML-file and this frequency will then become the one used in the adjustment
process.

If the parameter is not given, CFS defaults to the previous behaviour and
uses startFreq_ as base for the adjustment.