So far we always did a linear sampling of the frequency interval. Now it is also possible to do a logarithmic (linear after a log_10 transformation) or reverse logarithmic sampling.
Therefore a new method ComputeNextFrequency() was implemented that performs
the computation of actFreq_ in each step.

Added some consistency checks in the constructor concerning frequency interval
and number of frequencies, added a new attribute samplingType_ for storing
the sampling approach and removed saveType_ since it seems to be referenced
no where.