Changed the way we specify re-ordering. The default in the XML-file, if no 'reordering' attribute is specified in the 'matrix' element is to set the value to 'expertsChoice'.
The Expert() module was adapted to only change the reordering strategy, if
it is allowed to do so via the parameter allowChangeOfReordering parameter.
The strategy for choosing a re-ordering remains the same.

With the new approach we can specify an arbitrary re-ordering in the XML-file
without having to override the expert. This is more convenient, since in the
latter case we have to specify a whole bunch of other parameters, too.