So far only the DirectCoupledPDE class deleted the permutation vector obtained from OLAS, once the re-ordering was incorporated into the equation number map. Now this is also done in the SinglePDE case. However, there is a slight conceptual change.
It is no longer the PDE that deletes the buffer. Instead the PDE only acts as
a transitter. We pass to ReorderMapping() not the address of the buffer, but
instead the address of the pointer and set this to NULL in the method. Thus,
the PDE effectively looses all knowledge about the permutation.

The equation data class then can do with the permutation array to its liking.
Currently all three classes delete the buffer after incorporating the re-
ordering for the sake of a low memory footprint.