The recent change that made StopProgress() log to std::cout instead of std::cerr revealed a conceptual weakness in the WriteInfo::Error() method.
The method did not call StopProgress() to finish the progress log, but instead
issued a 'FAILED' statement itself. Since this statement was not converted to
std::cout we had a sort of race condition between the flushing of cout and
cerr that led to error reports with the following structure:

> ++ Generating remaining output files ...                       OK
> ++ Reading in the mesh ...                                     OK
> FAILED
>
>
> ERROR:
> HarmonicDriver::HarmonicDriver:
>  Could not find a section <harmonic> with tag = 'anyTag' in newParam
>
> The error occurred in 'harmonicDriver.cc' on line 55.
> See .trace-file for trace dump of function call tree.
>
> ++ Generating driver ...                                       ==

This is of course a little bit mis-leading, since it was the generation of
the driver that failed and not the import of the mesh file. So I changed
the class a little bit:

1) Error() now calls StopProgress() with FALSE as success flag to terminate
   the logging, if on is in progress.

2) Error messages are now formatted OLAS-style and we use the new commandLine
   handler to output the complete name of the trace--file, if function tracing
   is enabled.

3) It is no longer optional to pass __FILE__ and __LINE__ to Error().

4) StopProgress now issues a 'WARNED', if a warning occured and we passed
   success = TRUE.