As was to be feared the removal of memory leaks by making more classes actually delete stuff in their destructors triggers lots of other destructors and since we happily pass around so many pointers this lead to an "invalid free" occuring in the destructor of BaseForm e.g. in piezo-direct simulations.
This was related to the fact that BaseForm tried to delete ptMaterial. A
pointer to a material data object inserted into the form via the SetMaterial()
method by the PiezoCoupling object, which in turn got the information from
the MaterialData vector in BasePairCoupling, which in turn obtained the
information by passing a reference to the GetMaterialMethod() of
LoadMaterialDataFile.

The fact now is that the destructor of BasePairCoupling will delete all
its material data objects by the simple fact that at the end of
BasePairCoupling's life time the materialData_ StdVector will go out of
scope and its destructor will do the job. Requiescat in Pacem MaterialData.

Since there is no easy way to inform BaseForm on the sudden death of the
object ptMaterial points to. We do the following half-decent workaround:

BaseForm now has an attribute delMatDataAtEnd_. Only if this flag is true
will the destructor try to delete ptMaterial. The flag is set to true by
those two destructors that do a copy generation of a MaterialData object
and to false by the other two. (Actually who needs four constructors?)

To be sure also the SetMaterial() method sets delMatDataAtEnd_ = false. And
to be very sure it checks that ptMaterial is NULL before over-writing it
and deletes the old data if delMatDataAtEnd_ is true.