The class TimeStepping got a new method AdvanceTimestep, which in the case of fractional damping (NewmarkFracDamp) is responsible for storing past values to memory.
So Gerhard will implement a similar method in NewmarkFracDampMech.

By the way, in timestepping.cc a semicolon was missing in the destructor...

In order that the fractional damping works in nonlinear and/or
iteratively coupled PDEs, I moved the corrector to StepTransLin.
Also the counter for iteratively coupled PDEs is situated there.
PostStepTrans only contains the method AdvanceTimestep.
In IterSolveStep I moved the method PostStepTrans out of the iteration
loop, because it has to be excecuted only once at the end of the timestep.
I was not sure about IterSolveStep::SolveStepStatic and did not change the
position of PostStepStatic.
Also it does not seem logical to me, why in SolveStepStatic PostProcess is
excecuted and in SolveStepTrans we do not excecute it?