Major change in the way material tensors are treated:
Up to now, all material sub-tensors for axi, planeStrain and planeStress computations were taken from the yz-plane, because all piezoelectric materials are defined in the material file in such a way, that the polarization points in z-direction. 

In the new concept, the material in the material file is defined in an own coordinate system, denoted by the components (1,2,3). By default, the three components are now associated with the global directions (x,y,z). Also in the 2D-case, only the (1,2) components are taken. This is why in all 2D problems (axi, planeStress, planeStrain), where the polarization is assumed to be in y-direction, the material has now to be rotated around the x-axis by 90 degree. -> This was also done in Testsuite. Please refer e.g. to the examples coupledfield/piezo-direct/cube2d and /cubeaxi)

- This is why a general rotation mechanism for material is now available, which uses the Kardan angles to define arbitrary rotations. To rotate e.g. a material by 90 degree around the x-axis, one would define the following:

 <domain>
   <region name="bar" material="PZT">
     <rotation alpha="90" beta="0" gamma="0"/>
   </region>
 </domain>

  For further details about the rotation angles, please refer to the very detailed description in the xmlReference in section 4.8ff (Material assignment/coordinate systems).

- Added also a new local type of coordinate system "polar", which allows one to define an arbitrary polar coordinate system (only in 2D!)

- Unified the component names for the radial and cylindric coordinate systems:

  Cylindric coordinate system: (r,phi,z)
  Polar coordinate system:     (r,phi)

- The definition of a cylindric / polar coordinate system requires now the definition of the z/r-direction, instead of a point lying on this axis.