DRK may have a different opinion, but frequency is not equal to omega!
-> adapted PMLInt.cc

Now LinNeumannInt gets muparser object from baseform.

Implemented fractional damping in harmonic computation, that meant
changes to massInt.* baseForm.hh and acousticPDE.cc.
We evaluate a multiplicative prefactor for the mass integrator with 
the muparser, a habit others should follow.
The method is called SetExpression( str ), which passes the string
to the integrator.