This is the CFS part of the ilupack preconditioner/solver integration from OLAS (same date:))
The documentation is still pending and to come soon.
The quickest way to use ilupack is, to choose the solver
<solver type="ilupack"> and the rest is set automatically.

A more advanced example is

<solver type="ilupack">
  <ilupack>
    <permutation type="initial" matrix="sym" ordering="mmd"
                 matching="mwm"/>
    <permutation type="regular" matrix="sym" ordering="mmd" 
                 matching="mwm"/>
    <permutation type="final"   matrix="pd"  ordering="pp"
                 matching="none"/> 
    <engine type="sqmr"/>
    <elbowSpace>100</elbowSpace>
    <flag name="AggressiveDropping" state="off"/>
  </ilupack> 
</solver>

In any case check the .las file, the ilupack integration is 
quite a chatterbox.

Also the export of linear systems is extended, e.g. 
<exportLinSys baseName="mech2d" solution="yes" format="harwell-boeing"/>

Documentation to come.

Ilupack makes most sense with pardiso available. First tests show
a really boost on singlefield problems (e.g. from 72min to 5 min!)
But coupled fields either crash or give wrong results w/o warning!!

When exporting the solution ("exclusive" only writes the solution and
not the system) you can compare with a pardiso reference. In the las
file also the L2 norm of the solution is written on export.