The methods of boosts' filesystem library expect as arguments normall the canonical (platform-independent) representation of files and paths.
If we want have a native (platform-dependent) representation, we have to create a new boost::filsystem::path object, with the additional argument 'native' to indicate, that this is the native represenatioon.

Therefore the call to the method boost::filesystem::exixts within the constructor of Xerces now takes as argument a previously created path-object. 

This fixes the problem, that previously an error was thrown, if the schema or the xml file were located in a directory containing the literals 'CFS++', as this represents no valid canonical path, but instead a native representation.