It is now possible for almost all PDEs to write the RHS vector as a result into the gmv/gid/hdf5/unv file. Therefore we have now the following new nodal result types:
  * mechRhsLoad
  * elecRhsLoad
  * acouRhsLoad
  * heatRhsLoad
  * magRhsLoad

To write for example the RHS of a mechanic simulation (coming e.g. from a nodal load, a pressure integrator or from coupling quantities), we can specify

  <nodeResult type="mechRhsLoad">
    <allRegions/>
  </nodeResult>

NOTE: This feature works only for linear static, transient and harmonic simulations but NOT for non-linear static/transient ones (due to incremental solution).