Fixed a major bug:
In case the pardiso solver was used for piezo-electric problems, sometimes wrong results were computed.

This is now fixed, as we now enable some additional pardiso-internal scaling and matching strategy in the case we have a non-spd matrix (default case). For further information read the comment in the code or refer to the pardiso 3.1 user documentation for the parameters IPARM(11) and IPARM(13).