Reimplement Rayleigh Damping. Rayleigh damping can now be specified by alpha, beta and measuredFreq: <mechanicalDamping>   <rayleigh>     <alpha>     1.0e+3 </alpha>     <beta>      2.0e-6 </beta>     <measuredFreq>     1000.0 </measuredFreq>   </rayleigh> </mechanicalDamping>
or by lossTangensDelta and measuredFreq:
<mechanicalDamping>
  <rayleigh>
    <lossTangensDelta> 0.1 </lossTangensDelta>
    <measuredFreq>     1000.0 </measuredFreq>
  </rayleigh>
</mechanicalDamping>

The damping element rayleigh in the simulation xml file
has two new elements freq and ratioDeltaF:
In a transient simulation alpha and beta can be adjusted to the frequency freq
ratioDeltaF is the ratio of deltaF/freq. With this ratio the frequencies f1 and f2 are computed.
With f1, f2, tanDelta1, and tanDelta2 alpha and beta are finally computed.
The default value of ratioDeltaF=0.01

f1=freq*(1+ratioDeltaF)
f2=freq*(1-ratioDeltaF)

In a harmonic simulation freq is ignored.

<dampingList>
  <rayleigh id="1">
    <freq> 1000 </freq>
    <ratioDeltaF> 5.0 </ratioDeltaF>
  </rayleigh>
</dampingList>


[fixes #15]