mplemented the selection of arbitrary nodes / elements from the grid to define new namedNodes / namedElements. Before this check in, there was only a possibility to select nodes from a cylindrical coordinate system using the <directivityNodes> section.
This is now generalized. It is now possible to

- select one node / element by given coordinate and create a new namedNode/Element

- select a list of nodes, using a parametric description w.r.t. to a given coordinate system. This allows us e.g. to select nodes for directivity patterns or to select nodes lying on a line etc.

To select a single node with given coordinates and give it the name "myHist", simply adjust the <nodes> element within the <nodeList> section as follows:

    <nodeList>
      <nodes name="myHist">
        <coord x="1" y="1"/>
      </nodes>
    </nodeList>

To select a list of nodes on a given radius of 2m from the origin in a range from 0 deg to 180 deg in steps of 2 deg w.r.t. to a polar coordinate sysem with name "mid", we can write e.g.

   <nodeList>
     <nodes name="directivityNodes">
       <list coordSysId="mid">
         <freeCoord comp="phi" start="0" stop="180" inc="2"/>
         <fixedCoord comp="r" value="2"/>
       </list>
     </nodes>
   </nodeList>


Note, that these definitions also work in the <elemList>-section. There the element midpoint is used to determine the coordinate.

[closes #9]