Major change in the way we handle boundary conditions:
CFS allows now to prescribe loads, boundary conditions and constraints also on boundary elements / surface regions. Previously it was only possible to use namedNodes. This extension is necessary, as we want to be able in the future to map boundary conditions and couplings also to elements of higher order, where in fact a boundary condition / coupling is not only defined on nodes, but also on edges / surfaces.

This change induced some changes in OLAS as well. Most notably, the parameter bcNum is gone in the method 'SetDirichlet()', as the idbc-handler now takes track of the indices internally via a std::map.

FInally also the parameters referencing a certaing dof in block systems are gone within OLAS, which were e.g. still present in the methods:

- SetNodeRHS()
- AddToDiagMatrixEntry()
- SetDirichlet()
- Set/GetMatrixEntry()

The changes are verified against the TESTSUITE.