With student help a material file found his way into the repository. But...
Be aware!
Student helper was not able to verify xml-structure for all materials.
Also had problems with exponent of values. So I corrected the density of air
from 1290 kg/m^3 to 1,29 kg/m^3...
Some nonsense could still be hidden!

Anyway, it will only be found, if people use it!!

Cheers,
LB