Implemented new features to the HDF5 input reader.
This first rough version allows the user to query e.g. the number of analysis steps, type of analysis, defined results and their values. It implements an interface, which can read in a specific result. This result object can be used within the simulation itself (e.g. as rhs values, dirichlet values).