Fixed a bug:
The calculation of results using the postProc-methods for results, which calculate the sum/max/min over time/frequency, did not work. This is due to the fact, that the final writing of these results was performed after the related multisequence step was finished.

Replaced also some calls to SolutionType Enum2String(), which may throw exceptions, although one could use directly the resultName instead of the Enum-representation.