Fixed a minor bug:
In the hdf5 output writer some groups, which were opened in the method BeginMultiSequenceStep(), were closed in the destructor. This caused problems in the case when only a grid gets written, because the desctructor tries to close hdf5 groups, which are not open.

Not these groups are closed in the corresponding method FinishMultiSequenceStep()