Fixed a minor bug:
The selection mechanism for user-defined elements (select element by coordinate) did not check the dimension of the selected element. In the case of a mesh with volume and surface elements, the algorithm therefore sometimes pickes the surface element. However, this is mostly not wanted.
Therefore the algorithm now checks, if the selected element is a surface element and assigns to it a "penalty" distance, so that it will not be selected as nearest element to the selected coordinate.