This check in updates the handling of the SBM system:
- The xml parameters for the sbm system are now read in correctly

- Added copy-method for vector in the generatematvec.hh/.cc file

- Implemented dummy ID-preconditioner for sbm-system 

- Fixed a bug in the method CompRes in the sbm-matrix

- Fixed a bug in the idbchandlerpenalty

- Added missing copy and assignment operators for the sbm-vector

At the moment we can use the svm system to solve a coupled system using a conjugate gradient solver with identity preconditioner. However, the solution vector will not have the correct order, as the solvestep driver and the direct coupled pde do not really respect the pde id when querying the solution vector.