Commented out size check for assignment operator=():
Up to now a check for non-empty matrices was performed in the assignment
operator Matrix::operator=(), i.e. in the DEBUG-version it was not valid to
assign an empty matrix to another. However, exactly this happens in the file
ShapeDesign.cc:39, where an empty matrix gets push-backed. This caused an
unhandled exception, which prevented the successfull run of many optimization
test cases.

The size checks within the other operators (+,-,*,/ etc.) are still present, as
the calculation with zero-size matrices is mostly not meaningful.