Up to now GiD was not able to show all eigenmodes, if two or more eigenmodes are located at the same frequency.
Therefore we introduce now a grouping of eigenfrequencies by incorporating
explicitly the multiplicity of the eigenmode in the resultname like

  ef1//mechDisplacement
  ef1//mechStress
  ef2//mechDisplacement
  ef2//mechStress

for each frequency. This is shown within GiD in a menu-like structure

 ef1+-mechDisplacement
    |-mechStress
 ef2+-mechDisplacement
    |-mechStress

The grouping of eigenfrequencies can be switched by the attribute "groupEigenFrequencies" 
of the GiD output writer, which is by default "true".

  <fileFormats>
    <gid groupEigenFreqs="true"/>
  </fileFormats>