- Added finite element(s) and  -space for hierarchical H1 elements of   higher order. Currently the Legendre polynomials are hard-coded, but a future implementation will allow a variable choice of hierarchic shape functions.
- The FeSpaces have at the moment two possible implementations for
  handling hierarchical elements: One is to re-use the concept of
"nodes", i.e. we also introduce edge/face/inner nodes as in the Lagrange
case. However, we do not have to permute the nodes, as the orientation
is incorporated in the calculation of the shape functions itself. This
method is currently the favored one.
The second approach is the same as in the old EqnMap: We introduce new
arrays for edges, faces and interior mapping. This implementation is not
completed yet and will most likely be removed.

- The method PrintEqnMap() is now much more detailed.

- The passing of surface elements to the FeFunction-class was not
  properly performed in the SinglePDE-class.

- The element shape class now can handle surface elements properly and
  re-uses the reference elements instead of creating them each time the
SetElem() method is called.

- The ElementEntityType is now moved to BaseFE::EntityType.

- Information about approximation orders is now moved to the FeSpace
  class.