Fixed a bug in the determination of the propgation region for PML damping:
Up to now, the min/max extension of the "inner" propagation region was
initialized to 0. However, if the whole setup has a positive offset, the 
propagation region always started at x/y/z = 0, which can not be assumed
in general, leading to wrong results of the PML damping region!

Now we initalize the minimum of the inner region with a very large value 
and the maximum with a very small value, so that the min/max coordinate
of the propagation regions is determined correctly.

This also eliminates the ambigous case, when there is just a PML region
without propagation region present (e.g. testsuite example 
solsver/pardiso/mechAcoChannelPressurePML): In this case a unique set of
damping parameters can not be determined automatically. If this occurs, an exception is
thrown and the user has to specify the propagation region by hand using the
<propRegiopn> sub-element.