Introduced a clean separation of generation of error / warning and printing them:
The Exception class has now two additional methods: SetCallbackEx() and
SetCallbackEarn(), which both expect a callback function. This allows us 
to keep the exception class free of references to the ParamNode class or
the coloredConsole. Instead, we provide a callback function in the 
CFS class, which prints warnings to the console and the info.xml.