Attempt to fix the icc compile problem with BaseOptimizer.cc
AFAIK this is due to a intel bug. Currently the workaround is done within CFS 
in the code and in cmake files. This should clearly be unified but I'm not
sure where is the better place.

If it is a intel bug I would say it should NOT go to the code.

Needs to decided when Bastian's patch Bastian_Schmidt_MatOpt_0709@10237 goes
trunk.