- The Grid class has now an additional method, to write information about elements as pseudo results to the output writer.
Currently, the global representation of the element local coordinate directions
(xi, eta, zeta) are generated as vector results, as well as the Jacobian
determinant as scalar result. The result is generated in a separate
multisequence step with index 0.

To activate the grid information writing, the optional attribute "printGridInfo"
has to be set:

  <domain printGridInfo="yes" geometryType="3d">
     ...
   </domain>
   

-  Added a new section "variableList" within the <domain> element, which allows
to define additional mathParser variables with constant values. This is useful,
e.g. for transient simulation where several things (time step size, excitation)
depend on the frequency:

  <domain>
    <variableList>
      <var name="freq" value="1e6"/>
    </variableList>
    ...
  </domain>
  ....
  <transient>
    <deltaT> 1/freq/20 </deltaT>
    ....
  </transient>
  ...
  
   <load value="sin(2*pi*freq*t)"/>
   ....
Currently, the "value" of a variable can just be a constant and not an
expression itself

- Added a global Function "Transpose()" which can be used with our Matrix
class. In case of expression templates, this just reverses the index order and
thus has no runtime penalty. In case of debug compilation, this is a free
function, which returns a copy. This allows us to write expression like this

   globVec = Transpose(jacInv) * actLocDir;
   
i.e. one does not need to use the method "MultT" anymore or generate a copy by
calling the Matrix::Transpose() method.

- In the method BaseFE::Local2GlobalCoord(), the call to Matrix::Mult() got
replaced by a plain *-operator, as the Mult method is just defined on the
DenseMatrix class and thus has to perform a costly dynamic_cast operation.