Added method of Biot-Savart to calculate the fundamental solution of the magnetic field for "current sticks". Thus we do not have to physically "mesh" those coils but provide the coil layout by means of simple text files.
The law of Biot-Savart allows us to calculate analytically the magnetic field in
air of a given setup at any point in air. In our implementation the simple case
of "current sticks" is implemented, i.e. infinitely thin, straight lines, which
carry a known current value.

To define a coil within the magnetic PDE, one has to define e.g.

  <bcsAndLoads>
    ...
  </bcsAndLoads>
  
  <biotSavart>
    <coil name="myCoil" file="coilDescr.txt" value="1"/>
    <coil .../>
  </biotSavart>
  
Here "name" is just an arbitrary name (no region name!), "value" is the current
value in Ampere and "file" is the name of the file with the coil definition. 
This file defines the start/endpoint coordinates of the current sticks. These
coils are also visualized as 1D line elements in the postprocessing mesh.

The current implementation just works with the <magneticPDE> for static and
harmonic analysis, but not yet transient.

In this case, the fundamental solution for the vector potential is calculated 
once for each coil and gets added to the solution of the algebraic system in 
the method StdSolveStep::PostStepStatic() and ::PostStepTransient().

A testsuite example will be added in the next commit.