Fixed a bug regarding frequency dependent material parameters:
The first time the frequency is registered at the math parser, the
callback functions are not called yet and therefore frequency dependent
material parameters still use f=0 as frequency value.

Thus we initialize now the frequency variable
"f" to 0 in the constructor explicitly, so when the frequency loop is entered 
the first time, the callback functions get triggered and the materials are 
intialized properly.