Added new attribute "entityNumbering" to the <text> output writer.
This allows us to switch the naming scheme for history files:
In case of "global" numbering, the global node / element number is used
in the name of the history files. In case of "consecutive", each node / 
element of a list is numbered consecutively from 1, independent of 
the global number. This is advantageous, if the name of the history
files should be independent of the mesh discretization.