BUG in ReadRestart repaired
If <coord .../> is used to add nodes, it throughs an error due duplicate
adding of the nodes. This is now circumvented by reading in the grid of
the restart file but then overwriting it again with a copy of the
original grid.

There is no check that these even match--still the responsibility of the
user.