new calc Rhs Load and clean up --
Calculating the Lighthill Tensor (acouRhsLoad) from fluid flow which
uses quadratic elements causes numerical errors which result in
fluctuations in the acouRhsLoad. To be precise, corner nodes have the
oposite sign as edge nodes.
As a remdy a new routine has been implemented which calculates the
Lighthill Tensor at the centre of an element and is then distributes it
to the neighbouring nodes. Since each nodes has different element
neighbours, each node will still have a different value.
The call for this option is:
$> cplreader --doIntAverageCentre 1 [...]

Furthermore, the standard method has been cleaned up for better
readability.

linearForm.*:
  - new method for calculating acouRhsLoad by just regarding the centre
    of an element
  - clean up of standard acouRhsLoad calculation
cplreader_options:
  - new option to calculate the acouRhsLoad using the centre of an
    element (--doIntAverageCentre 1)
  - missing option for outputfield were added
elemIntegr.*:
  - new call for the new method
CouplingHandler.*:
  -new flag needs to be checked to call the new method