stress/strain and von mises stress/strain are now evaluated at the integration points instead of only the center point.
For a regular grid stress/strain is still the same but the von mises norm changes.

Also the density can evaluates by the attribute stress_strain_density.

This is also reflected in stress (stressDensity) topology optimization.

The piezostress is in postproc still a the center point.

Testcases are not updated, yet