make named nodes from lines --
gmsh can not create named nodes from lines so cfs++ needs to do it.
This can be specified in surfReginList. Nodes on a second line may be
told to be exluded. This may be needed for instance for inflow boundary
condition were only the inner nodes of a line are to have a value and
not those on the wall.
The new entity gets the suffix "nodes_" and may then be used by coupling
conditions, boundary conditions etc.