Major changes due to merge r9147-11125 from trunk.
We now have the current version of the ParamNode, as well as I/O classes 
methods available in this branch.

No new functionality is available compared to the previous version.

Optimization is still completely deactivated. Only the electrostatic PDE 
was tested with the condaxi example. 

All occurrences of INGEGLIB and MEMTRACE have been removed, as well as 
the generation of .deb and .mem files. This functionality should
be covered by more advanced libraries (e.g. google perftools, likwid) in 
the future.