repaired Centos Bug --
g++ on Centos was complaining:
/.../CFS_TRUNK_NIGHTLY/source/PDE/eqnMap.cc:1421: warning: dereferencing
type-punned pointer will break strict-aliasing rules


So the new methods were enhanced to templates, hence not casting to
ResultHdBcMap anymore. ResultHdBcMap is a map of vector, which stored
the boundary conditions. This is quite complex and casting this whas
quite dangerous. The template way is cleaner.