his is the first check in regarding the novel interator concept. NOTE: this is more or less a preliminary check in to anable the coworkers to get used to the new integrator definition scheme and enable the developement of own ideas
INtsted of defining for each pde a specialized integrator we have
now a BDB-Int and BBInt class.
The later is just for integrators which do not have a material
tensor.
both classes get the required operators as template parameters.
Therefore the complexity of a specific
form is moved to the Operator.
Defined operators for 
 - identity
 - gradient
 - curl

The Newly definied operators are listed below. Some files like
the orignial bdbInt.hh were renamed to bdbInt.hh_obsolete
One should take care which files are now in charge and which
are not. 
In the next checkin, the old files which are no longer
needed will be deleted

There are some major issues in the current implementation:
  1. The mass integrator for magEdgePDE had to be implemented within a specialized class due to the fact, that 
     The solution strategy requires a call to SetLowestOrder of FeHCurlHi This should be changed in the future
  2. All nonlinearities in MagEdgePDE were commented out. The coefFunction concept which is to be developed will reenable them
  3. There is no replacement for the LinearForm class. This will be the RHSInt class which will be implemented next week.
  4. For debugging one has to take care that he uses a more sophisitcated debugger than gdb (e.g.  totalview). Templates are not 
     nicely supported by gdb.
  5. The implementation features some part of code which are not tested but prepared for the time they will be used