NOTE: This is the first checking, to be tested if everything works for a clean checkout... so be prepared that the next checking will be done soon...
Resturced the source directory, renamed every file according to the new
convention and implemented a right hand side
integrator BUInt for electrostatics

The new directory structure is made as specified last friday.
Unfortunately the cmake structure is a little messed up right now due to
the major changes. So this has to be cleaned up. But at least the code
should compile and link correctly. 

Next Steps are:
 1. Documentation of the directory structure in the Manuals
 2. Clean up cmake dependencies
 3. Make changes according to suggestions
 4. Refill the file headers with the coding conventions rgarding the
    tabs etc. Also we include a standard comment header in each hh
    file.
 5. Proceed with everything :)