This commit is an intermediate one, mainly to transfer the current code version to another workstation. Sorry for this.
This revision provides functionality but
disregards documentation in too many places. Therefore there will be a
documentation and code comments checkin tomorrow.

Anyway, this is the very first version of the new time stepping scheme
concept as given in the develManual.
Works now for heatPDE and acousticPDE.

It allows the definition of timescheme per FeFunction and drops
notations such as EffectiveMass or Stiffness. 
Most importantly it is now an important convention that the matrices are
associated to a certain time derivative order
mass -> second order time derivative
damping -> first order time derivative
stiffness -> zero order time derivative

EVERY pde should respect this convetion. So the heatPDE now has a damping
and a stiffeness matrix and no mass matrix.

Already removed the "old" files newmark, trapezoidal, etc.
this has been done to avoid confusion. Perhaps we will need some of
these classes later when it comes to more sophisticated schemes.