Fixed two bugs.
First the includes in FlowMaterial.cc were wrongly spelled
Second, the PreCalcShapeFnc of FeSpaceH1Nodal caused a fault if the user
specifies just an integration scheme. Right now, this is fixed by an
additional map in the FeSpace class which should now allow arbitrary
combinations of approximation order and integration schemes.

ONE BIG TODO> Right now, the GetIntegrationPoints method throws an
exception if a specific integration order has no points. This has to be
changed. Its no use to throw an exception if no intergation points are
available for order 8 and there are only quads in the grid...