Corrected some errors with transient mixed formulations.
Furthermore the FeFunction is now capable to write out results on
quadratic grids. Computation of arbitrary orders on quadratic grid was
possible anyway
To accomplish this, the BaseBOperator interface was
changed and the FETYPE template parameter was removed from this base
class as it comes to use only in the derived classes.