Fixed a problem regarding the assembly of free->idbc entries of transposed blocks.
In case an off-diagonal matrix block has entries idbc->free and also the transposed
block has to be set (free->idbc), those entries hav not been assembled yet due to
a missing / buggy if-case in the AlgebraicSys::SetElementMatrix() and 
GraphManger::SetElementPos(). This error only ocurred when using the 
"elimination" approach for idbc values.

By fixing this, the mixed acousticPDE now also works when assembling the K_PV
block (instead of the K_VP block before).