This checkin reenables the PML for acousticPDE in case of harmonic simulations. The CoeffunctionPML is already prepared to handle also the transient cases but its not used right now. The GetScalar method of PMLCoefFunction returns the mutliplied values of all damppingcoefficients and is thereby used to transform the jacobian. The clean way would be to transform the jacobian directly in the pml regions, i.e. in the Local point mapped class. But as this would cause other problems we choose the mentioned path here. The transform jacobian mehtod of the operators is thereby obsolete. At least for the PML functionality.
For the transformation of the jacobian, we use the newly defined
CoefFunction Compound. The mixture of complex and real coefficient
functions worked out of the box (hurray for andi hauck) with one
exception:
The imagniary variable always gets a name in
CoefFunction::GenScalCompNames regardless of their datatype. Otherwise
there would be an error later when the Compound function checks for the
validity of variable names.

To take damping values into account the ScaledGradientOperator is
introduced. It is basically the standard gradient which multiplied the
vector coeffunction to the spatial derivatives.

Right now the PML prints the computed bounding box of propagation region and pml
region to the terminal to give the user a better insight to implicitly
computed geometrical definitions. This output will  be transfered to
info-xml in future.